Solve the problem of number of video limitation
I have uploaded the file in our server, do need to change the file in server any more.

+++ converter.php
@@ -1,7 +1,7 @@
 <html>
 <body>
 
-<?php
+<?php 
 if(isSet($_GET['id'])){
   $videoName = $_GET['id'];
   echo $videoName;
@@ -9,21 +9,28 @@ if(isSet($_GET['id'])){
 
 //$str = file_get_contents("/var/www/html/web2016/owen_test/php/test_data.json");
 //$json = json_decode($str, true);
-
-$dir = "./temp_video/".$videoName;
-echo $dir;
+echo $_File_;
+$dir = "/var/www/html/web2017/project/temp_video";
 if( !file_exists($dir)) {
   echo "in";
-  $oldmask = umask(0);  // helpful when used in linux server
+  $oldmask = umask(0);  // helpful when used in linux server  
   $res = mkdir($dir, 0744);
   echo $res;
 }
-
 else {
   $files = glob($dir."/*"); // get all file names
+  $sum = 0;
   foreach($files as $file){ // iterate files
-  if(is_file($file))
-    unlink($file); // delete file
+    if(is_file($file)) {
+      $sum++;
+    }
+  }
+  foreach($files as $file){
+    if ($sum >= 8) {
+      if(is_file($file)){
+        unlink($file); // delete file
+      }
+    }
   }
 }
 
@@ -33,7 +40,7 @@ $newvideoName = substr($videoName, 0, -4);
 $newvideoFile = $dir."/".$newvideoName.".mp4";
 $cmd = "$ffmpeg -i $videoFile $newvideoFile";
 shell_exec($cmd);
-
+//echo("<script>console.log('".shell_exec($cmd)."');</script>");
 echo "Load video successfully!";
 ?>
